requirements.txt
flask
python-dotenv
pillow
requests
runwayml
openai
torch
torchvision
timm
scikit-learn
pandas
opencv-python
matplotlib